Migrate user comments to posts reducer
