+++ src/comments/commentsReducer.js
@@ -1,5 +1,4 @@
 import * as commentsTypes from './commentsActions';
-import * as userTypes from '../user/userActions';
 
 const initialState = {
   childrenById: {},
@@ -43,17 +42,6 @@ const commentsData = (state = {}, action) => {
         ...state,
         ...mapCommentsBasedOnId(action.payload.content, action),
       };
-    case userTypes.GET_USER_COMMENTS_SUCCESS:
-    case userTypes.GET_MORE_USER_COMMENTS_SUCCESS: {
-      const commentsMoreList = {};
-      action.payload.forEach((comment) => {
-        commentsMoreList[comment.id] = comment;
-      });
-      return {
-        ...state,
-        ...commentsMoreList,
-      };
-    }
     case commentsTypes.RELOAD_EXISTING_COMMENT:
       return {
         ...state,
@@ -81,8 +69,6 @@ const comments = (state = initialState, action) => {
     case commentsTypes.GET_COMMENTS_START:
     case commentsTypes.GET_COMMENTS_SUCCESS:
     case commentsTypes.GET_COMMENTS_ERROR:
-    case userTypes.GET_USER_COMMENTS_SUCCESS:
-    case userTypes.GET_MORE_USER_COMMENTS_SUCCESS:
       return {
         ...state,
         comments: commentsData(state.comments, action),
